+++ src/main/java/com/epam/repository/TrainerRepository.java
@@ -3,6 +3,8 @@ package com.epam.repository;
 import com.epam.entity.TrainerEntity;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public class TrainerRepository extends MapRepositoryImpl<TrainerEntity, Long> {
+public class TrainerRepository extends MapRepositoryImpl<TrainerEntity, BigInteger> {
 }
