+++ src/main/java/com/epam/repository/TrainingRepository.java
@@ -3,6 +3,8 @@ package com.epam.repository;
 import com.epam.entity.TrainingEntity;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public class TrainingRepository extends MapRepositoryImpl<TrainingEntity, Long> {
+public class TrainingRepository extends MapRepositoryImpl<TrainingEntity, BigInteger> {
 }
