+++ src/main/java/com/epam/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.epam.repository;
 import com.epam.entity.UserEntity;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public class UserRepository extends MapRepositoryImpl<UserEntity, Long> {
+public class UserRepository extends MapRepositoryImpl<UserEntity, BigInteger> {
 }
